+++ src/components/layout/customer/components/write-offs/modals-content/EditResponseModalContent.tsx
@@ -1,0 +1,14 @@
+import QuestionnairesStepper from '@/components/QuestionnairesStepper';
+import { useState } from 'react';
+
+const EditResponseModalContent = () => {
+  const [currentStepIndex, setCurrentStepIndex] = useState(0);
+  return (
+    <QuestionnairesStepper
+      currentStepIndex={currentStepIndex}
+      setCurrentStepIndex={setCurrentStepIndex}
+    />
+  );
+};
+
+export default EditResponseModalContent;
